add conanfile.py, and fix a couple windows warnings.

+++ parallel_hashmap/phmap.h
@@ -3561,8 +3561,8 @@ namespace memory_internal {
 // ----------------------------------------------------------------------------
 template <class Pair, class = std::true_type>
 struct OffsetOf {
-  static constexpr size_t kFirst = -1;
-  static constexpr size_t kSecond = -1;
+    static constexpr size_t kFirst = (size_t)-1;
+  static constexpr size_t kSecond = (size_t)-1;
 };
 
 template <class Pair>
